# feat: add new feature to disable compression for parquet on ingester