--- a/src/cli/basic/mod.rs
+++ b/src/cli/basic/mod.rs
@@ -16,3 +16,4 @@
 pub mod cli;
 mod http;
 mod load;
+mod test;