--- a/src/config/src/meta/stream.rs
+++ b/src/config/src/meta/stream.rs
@@ -263,6 +263,7 @@ impl From<&String> for QueryPartitionStrategy {
 pub enum MergeStrategy {
     FileSize,
     FileTime,
+    TimeRange,
 }
 
 impl From<&String> for MergeStrategy {
@@ -270,6 +271,7 @@ impl From<&String> for MergeStrategy {
         match s.to_lowercase().as_str() {
             "file_size" => MergeStrategy::FileSize,
             "file_time" => MergeStrategy::FileTime,
+            "time_range" => MergeStrategy::TimeRange,
             _ => MergeStrategy::FileSize,
         }
     }