--- a/src/infra/src/file_list/postgres.rs
+++ b/src/infra/src/file_list/postgres.rs
@@ -1220,10 +1220,13 @@ SELECT stream, max(id) as id, COUNT(*)::BIGINT AS num
         DB_QUERY_NUMS
             .with_label_values(&["update", "file_list_jobs"])
             .inc();
-        let sql = format!(r#"UPDATE file_list_jobs SET updated_at = $1 WHERE id IN ({})"#, ids.iter()
+        let sql = format!(
+            r#"UPDATE file_list_jobs SET updated_at = $1 WHERE id IN ({})"#,
+            ids.iter()
                 .map(|id| id.to_string())
                 .collect::<Vec<_>>()
-                .join(","));
+                .join(",")
+        );
         sqlx::query(&sql)
             .bind(config::utils::time::now_micros())
             .execute(&pool)