- [x] Add time_range strategy for compactor
- [x] Add none-compression for ingester
- [x] Fix tracing for search

New ENV
```
ZO_FEATURE_INGESTER_NONE_COMPRESSION=false
```
We can disable compression for parquet on ingester if you enabled it.

New Options
```
ZO_COMPACT_STRATEGY = time_range
```
We can use `time_range` strategy to sort by files try to avoid time range overlapping which can merge files without `oder by _timestamp` let the merge faster.

## Conclusion

1. For write on ingester, almost same.
2. For query on ingester, around 6% performance improve
3. For compaction, almost same.

## TODO
add more testing result, What is the performance difference

- [x] Write speed on ingester
- [x] Search speed on ingester

## Write testing
Without compression

|Type   |  Name                   |                  # reqs  |      # fails |    Avg   |    Min   |    Max   |   Med |   req/s  |  failures/s | 
|-------|------------------|------------------|----------|--------|--------|--------|-------|--------|-----------|
|POST  |   /locust1/_multi   |               609412   |    0(0.00%) |     26  |         7 |     451  |       27 |  368.96 |        0.00 | 
| -         | Aggregated          |               609412   |    0(0.00%) |      26  |        7 |      451 |       27 |  368.96 |        0.00 |

Response time percentiles (approximated)
|Type     | Name                  |   50% | 66% |  75% |  80% |  90% |  95% |  98% |  99% | 99.9% | 99.99% |  100% | # reqs |
|--------|-----------------|-------|------|------|-------|------|-------|------|-------|-------|---------|--------|-------|
|POST    |  /locust1/_multi  |       27 |     29 |     30 |      31 |     34 |      38 |     45 |      51 |       91 |        130 |      450 | 609412 |
| -           | Aggregated        |       27 |     29 |     30 |      31 |     34 |      38 |     45 |      51 |       91 |        130 |      450 | 609412 |


With compression
|Type   |  Name                   |                  # reqs  |      # fails |    Avg   |    Min   |    Max   |   Med |   req/s  |  failures/s | 
|-------|------------------|------------------|----------|--------|--------|--------|-------|--------|-----------|
|POST  |   /locust1/_multi   |               641241   |    0(0.00%) |     26  |         8 |     165  |       27 |  366.56 |        0.00 | 
| -         | Aggregated          |               641241   |    0(0.00%) |      26  |        8 |      165 |       27 |  366.56 |        0.00 |

Response time percentiles (approximated)
|Type     | Name                  |   50% | 66% |  75% |  80% |  90% |  95% |  98% |  99% | 99.9% | 99.99% |  100% | # reqs |
|--------|-----------------|-------|------|------|-------|------|-------|------|-------|-------|---------|--------|-------|
|POST    |  /locust1/_multi  |       27 |     29 |     30 |      31 |     34 |      39 |     45 |      50 |       87 |        120 |      170 | 641241 |
| -           | Aggregated        |       27 |     29 |     30 |      31 |     34 |      39 |     45 |      50 |       87 |        120 |      170 | 641241 |

The write performance almost same, maybe because of the CPU usage didn't 100%, i have 12 CPU cores, only used half.

## Query testing
100GB compressed, 10 rounds
5.69s, 5.67s, 5.67s, 5.82s, 5.79s, 5.36s, 5.97s, 5.23s, 5.67s, 5.35s

100GB uncompressed, 10 rounds
5.27s, 5.13s, 5.47s, 5.28s, 5.04s, 5.31s, 5.12s, 5.34s, 5.22s, 5.22s

Summary:
Compressed data avg: 5.622s
Uncompressed data avg: 5.24s
Absolute improvement: 0.382s
Relative improvement: 6.79% faster with uncompressed data
The uncompressed data processing is about 6.79% faster than the compressed data processing.